+++ app/presenters/organisation_presenter.rb
@@ -2,10 +2,18 @@
 
 class OrganisationPresenter < SimpleDelegator
   def language_code
-    super.downcase
+    return I18n.t("organisation.language_code.#{super.downcase}") if super
+
+    super
   end
 
   def default_currency
-    super.downcase
+    return I18n.t("generic.default_currency.#{super.downcase}") if super
+
+    super
+  end
+
+  def organisation_type
+    I18n.t("organisation.organisation_type.#{super}")
   end
 end