+++ spec/features/staff/beis_users_can_view_organisations_spec.rb
@@ -66,6 +66,17 @@ RSpec.feature "BEIS users can view other organisations" do
     let!(:delivery_partner_organisations) { create_list(:delivery_partner_organisation, 3) }
     let!(:matched_effort_provider_organisations) { create_list(:matched_effort_provider, 2) }
     let!(:external_income_provider_organisations) { create_list(:external_income_provider, 2) }
+    let!(:implementing_organisations) do
+      2.times {
+        create(:implementing_organisation).tap do |org|
+          OrgParticipation.create(
+            organisation: org,
+            activity: create(:project_activity),
+            role: "implementing"
+          )
+        end
+      }
+    end
 
     before do
       authenticate!(user: user)
@@ -74,6 +85,22 @@ RSpec.feature "BEIS users can view other organisations" do
     context "when a role is provided" do
       before do
         visit organisations_path(role: role)
+      end
+
+      context "the role is 'implementing_organisations'" do
+        let(:role) { "implementing_organisations" }
+
+        scenario "lists all organisations in the 'implementing' scope" do
+          expect(Organisation.implementing.count).to be >= 2
+
+          within ".organisations" do
+            Organisation.implementing.each do |org|
+              expect(page).to have_content(org.name)
+            end
+          end
+          expect(page).to have_css(".organisation", count: Organisation.implementing.count)
+          then_breadcrumb_shows_type_of_organisation(name: "Implementing organisations")
+        end
       end
 
       context "the role is 'delivery_partners'" do
@@ -86,12 +113,20 @@ RSpec.feature "BEIS users can view other organisations" do
             click_on t("tabs.organisations.matched_effort_providers")
           end
 
+          it "includes type of organisation in breadcrumb" do
+            then_breadcrumb_shows_type_of_organisation(name: "Matched effort providers")
+          end
+
           include_examples "lists matched effort provider organisations"
         end
 
         context "when viewing the external income providers tab" do
           before do
             click_on t("tabs.organisations.external_income_providers")
+          end
+
+          it "includes type of organisation in breadcrumb" do
+            then_breadcrumb_shows_type_of_organisation(name: "External income providers")
           end
 
           include_examples "lists external income provider organisations"
@@ -149,6 +184,12 @@ RSpec.feature "BEIS users can view other organisations" do
       end
 
       include_examples "lists delivery partner organisations"
+    end
+
+    def then_breadcrumb_shows_type_of_organisation(name:)
+      within ".govuk-breadcrumbs" do
+        expect(page).to have_content(name)
+      end
     end
   end
 end